refactor[#73] : session 정보 Context 로 하나의 state로 관리
