+++ blog/app/@modal/(.)auth/page.tsx
@@ -3,12 +3,11 @@
 import { EmailLoginForm } from "@/features/auth/ui";
 import { LogoutForm } from "@/features/auth/ui/LogoutForm";
 
-import { useSession } from "@/entities/user/model/useSession";
-
+import { useSession } from "@/shared/model";
 import { RoutingModal } from "@/shared/ui/RoutingModal";
 
 const AuthModal = () => {
-  const { user } = useSession();
+  const user = useSession();
 
   return (
     <RoutingModal>{user ? <LogoutForm /> : <EmailLoginForm />}</RoutingModal>
