+++ blog/src/app/ServiceProvider.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { SessionProvider } from "./SessionProvider";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import React, { PropsWithChildren } from "react";
@@ -25,7 +26,7 @@ export const ServiceProvider: React.FC<PropsWithChildren> = ({ children }) => {
   return (
     <QueryClientProvider client={queryClient}>
       <ReactQueryDevtools initialIsOpen={false} />
-      {children}
+      <SessionProvider>{children}</SessionProvider>
     </QueryClientProvider>
   );
 };
