--- blog/src/entities/user/model/index.ts
@@ -1,1 +1,0 @@
-export * from "./useSession";
